fixing client
